# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Blog",
        "description": "My blog api documentation",
        "contact": {
            "email": "minicatdd7@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "v1"
        }
    ],
    "paths": {
        "/api/v1/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get list post",
                "description": "Return list of post",
                "operationId": "listPost",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Store a new post",
                "description": "Return a new post created",
                "operationId": "storePost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Store a new post successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/post/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get post detail",
                "description": "Return post detail",
                "operationId": "getPostDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found post"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete post",
                "description": "Delete any post you want",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted post"
                    },
                    "404": {
                        "description": "Not found post"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Post": {
                "title": "Post",
                "description": "Post model",
                "properties": {
                    "id": {
                        "type": "number",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "A nice post"
                    },
                    "content": {
                        "type": "string",
                        "example": "This is a content of post"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-26T14:24:23.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-26T14:24:23.000000Z"
                    },
                    "thumbnail": {
                        "type": "string",
                        "example": "https://via.placeholder.com/1080x1080.png/000055?text=cats+voluptate"
                    },
                    "summary": {
                        "type": "string",
                        "example": "This is a summary post"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            },
            "StorePostRequest": {
                "title": "Store post request",
                "description": "Store post request body data",
                "required": [
                    "title",
                    "summary",
                    "content"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "A nice post title"
                    },
                    "summary": {
                        "type": "string",
                        "example": "This is new summary on post"
                    },
                    "content": {
                        "type": "string",
                        "example": "This is nice post content"
                    },
                    "thumbnail": {
                        "type": "string",
                        "example": "https://via.placeholder.com/1080x1080.png/000055?text=cats+voluptate"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StorePostRequest"
                }
            },
            "PostResource": {
                "title": "PostResource",
                "description": "Post resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PostResource"
                }
            }
        }
    },
    "security": [
        []
    ]
}